Reflection based runner initializes file readers now
This also means that the Runner interface no longer provides an
addFileReader method for the problems to use.

The ReflectionBasedRunner iterates over all files in the "problemfiles"
directory and calls the addFileReader method based on the file name
(e.g. problem8number.txt is assigned to the alias "8" by regex matching)